+++ src/Clients/CommandableHttpClient.cs
@@ -79,7 +79,7 @@ namespace PipServices3.Rpc.Clients
         /// <param name="correlationId">(optional) transaction id to trace execution through call chain.</param>
         /// <param name="requestEntity">body object.</param>
         /// <returns>result of the command.</returns>
-        public async Task<T> CallCommandAsync<T>(string route, string correlationId, object requestEntity)
+        protected async Task<T> CallCommandAsync<T>(string route, string correlationId, object requestEntity)
             where T : class
         {
             var timing = Instrument(correlationId, _baseRoute + "." + route);