+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for .NET Changelog
 
-## <a name="3.3.12-3.3.13"></a> 3.3.12-3.3.13 (2020-12-28)
+## <a name="3.3.12-3.3.14"></a> 3.3.12-3.3.14 (2021-01-06)
 
 ### Features
 * Added rest route metadata to generate OpenApiSpec