+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for .NET Changelog
 
-## <a name="3.3.15-3.3.19"></a> 3.3.15-3.3.19 (2021-02-23)
+## <a name="3.3.15-3.3.20"></a> 3.3.15-3.3.20 (2021-02-23)
 
 ### Fixes
 * **clients** Fixed processing of empty routes
+* **clients** Fixed adding null filter/paging parameters
 
 ### Features
 * Extended RestClient to include CorrelationID into request headers