+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for .NET Changelog
+# <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for Pip.Services in .NET Changelog
+
+## <a name="3.4.0"></a> 3.4.0 (2021-04-22)
+
+### Fixes
+* **test** Added TestRpcClient
+* **test** Added TestCommandableHttpClient
 
 ## <a name="3.3.15-3.3.22"></a> 3.3.15-3.3.22 (2021-04-05)